Move AmazonCloudWatchModule to the messaging lib
And rename it to reflect the fact that most consumers really want the
MetricsSender it provides, not AmazonCloudWatch.
