+++ catalogue_pipeline/id_minter/src/main/scala/uk/ac/wellcome/platform/idminter/Server.scala
@@ -10,6 +10,7 @@ import com.twitter.finatra.http.filters.{
 import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.finatra.controllers.ManagementController
+import uk.ac.wellcome.messaging.metrics.MetricsSenderModule
 import uk.ac.wellcome.messaging.sns.{SNSClientModule, SNSConfigModule}
 import uk.ac.wellcome.messaging.sqs.{
   SQSClientModule,
@@ -32,7 +33,7 @@ class Server extends HttpServer {
     SQSReaderModule,
     SNSConfigModule,
     SNSClientModule,
-    AmazonCloudWatchModule
+    MetricsSenderModule
   )
 
   override def configureHttp(router: HttpRouter) {
