+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/transformer/Server.scala
@@ -10,6 +10,7 @@ import com.twitter.finatra.http.filters.{
 import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.finatra.modules._
+import uk.ac.wellcome.messaging.metrics.MetricsSenderModule
 import uk.ac.wellcome.messaging.sns.{SNSClientModule, SNSConfigModule}
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
 import uk.ac.wellcome.storage.dynamo.{DynamoClientModule, DynamoConfigModule}
@@ -21,7 +22,7 @@ object ServerMain extends Server
 class Server extends HttpServer {
   override val name = "uk.ac.wellcome.platform.transformer Transformer"
   override val modules = Seq(
-    AmazonCloudWatchModule,
+    MetricsSenderModule,
     AWSConfigModule,
     DynamoConfigModule,
     AkkaModule,
