+++ data_api/snapshot_generator/src/main/scala/uk/ac/wellcome/platform/snapshot_generator/Server.scala
@@ -12,6 +12,7 @@ import uk.ac.wellcome.display.modules.DisplayJacksonModule
 import uk.ac.wellcome.elasticsearch.finatra.modules.ElasticClientModule
 import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.finatra.modules._
+import uk.ac.wellcome.messaging.metrics.MetricsSenderModule
 import uk.ac.wellcome.messaging.sns.{SNSClientModule, SNSConfigModule}
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
 import uk.ac.wellcome.platform.snapshot_generator.modules.{
@@ -33,7 +34,7 @@ class Server extends HttpServer {
     "uk.ac.wellcome.platform.snapshot_generator SnapshotGenerator"
 
   override val modules = Seq(
-    AmazonCloudWatchModule,
+    MetricsSenderModule,
     AWSConfigModule,
     SQSClientModule,
     SQSConfigModule,
