+++ reindexer/reindex_worker/src/main/scala/uk/ac/wellcome/platform/reindex_worker/Server.scala
@@ -10,6 +10,7 @@ import com.twitter.finatra.http.filters.{
 import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.finatra.modules._
+import uk.ac.wellcome.messaging.metrics.MetricsSenderModule
 import uk.ac.wellcome.messaging.sns.{SNSClientModule, SNSConfigModule}
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
 import uk.ac.wellcome.platform.reindex_worker.modules.ReindexerWorkerModule
@@ -22,7 +23,7 @@ class Server extends HttpServer {
   override val name = "uk.ac.wellcome.platform.reindexer Reindexer"
 
   override val modules = Seq(
-    AmazonCloudWatchModule,
+    MetricsSenderModule,
     DynamoClientModule,
     DynamoConfigModule,
     SQSClientModule,
