+++ sierra_adapter/sierra_reader/src/main/scala/uk/ac/wellcome/platform/sierra_reader/Server.scala
@@ -10,6 +10,7 @@ import com.twitter.finatra.http.filters.{
 import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.finatra.controllers.ManagementController
+import uk.ac.wellcome.messaging.metrics.MetricsSenderModule
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
 import uk.ac.wellcome.platform.sierra_reader.modules.SierraReaderModule
 import uk.ac.wellcome.storage.s3.{S3ClientModule, S3ConfigModule}
@@ -21,7 +22,7 @@ class Server extends HttpServer {
     "uk.ac.wellcome.platform.sierra_reader SierraReader"
   override val modules = Seq(
     SierraReaderModule,
-    AmazonCloudWatchModule,
+    MetricsSenderModule,
     AWSConfigModule,
     S3ClientModule,
     S3ConfigModule,
